requirements.txt
aiofiles==23.2.1
aiogram==3.2.0
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
et-xmlfile==1.1.0
frozenlist==1.4.0
idna==3.6
magic-filter==1.0.12
multidict==6.0.4
numpy==1.26.2
openpyxl==3.1.2
pandas==2.1.0
pydantic==2.5.2
pydantic_core==2.14.5
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
soupsieve==2.5
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
xlrd==2.0.1
yarl==1.9.4
